+++ src/NServiceBus.Transport.SQS/RateLimiter/SizeConstrainedStack.cs
@@ -1,0 +1,27 @@
+﻿namespace NServiceBus.Transport.SQS
+{
+    using System.Collections.Generic;
+
+    partial class RateLimiter
+    {
+        class SizeConstrainedStack<T> : LinkedList<T>
+        {
+            public SizeConstrainedStack(int maxSize)
+            {
+                this.maxSize = maxSize;
+            }
+
+            public void Push(T item)
+            {
+                AddFirst(item);
+
+                if (Count > maxSize)
+                {
+                    RemoveLast();
+                }
+            }
+
+            readonly int maxSize;
+        }
+    }
+}
